Add option to keep closed channels

+++ src/ckb/channel.rs
@@ -185,14 +185,21 @@ pub struct ChannelActor<S> {
     peer_id: PeerId,
     network: ActorRef<NetworkActorMessage>,
     store: S,
+    keep_on_closed: bool,
 }
 
 impl<S> ChannelActor<S> {
-    pub fn new(peer_id: PeerId, network: ActorRef<NetworkActorMessage>, store: S) -> Self {
+    pub fn new(
+        peer_id: PeerId,
+        network: ActorRef<NetworkActorMessage>,
+        store: S,
+        keep_on_closed: bool,
+    ) -> Self {
         Self {
             peer_id,
             network,
             store,
+            keep_on_closed,
         }
     }
 
@@ -1229,7 +1236,9 @@ where
         match state.state {
             ChannelState::Closed => {
                 myself.stop(Some("ChannelClosed".to_string()));
-                self.store.delete_channel_actor_state(&state.get_id());
+                if self.keep_on_closed {
+                    self.store.delete_channel_actor_state(&state.get_id());
+                }
             }
             _ => {
                 self.store.insert_channel_actor_state(state.clone());
